Refactor data service and triple store implementations
